requirement.txt
dj-database-url==0.3.0
Django==1.7.8
django-apogee==0.0.0
django-classy-tags==0.6.1
django-cms==3.1.0
django-crispy-forms==1.4.0
django-reversion==1.8.5
django-sekizai==0.8.1
Django-Select2==4.3.1
django-treebeard==3.0
django-xadmin==0.5.0
djangocms-admin-style==0.2.7
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.7
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.5.2
djangocms-video==0.1
duck-examen==0.0.0
duck-inscription==0.1b0
duck-inscription-payzen==0.0.0
duck-paiement-etudiant==0.0.0
duck-salle==0.0.0
duck-site==0.0.0
duck-svi==0.0.0
duck-theme-ied==0.0.0
duck-utils==0.0.0
foad==0.0.0
gestion-info==0.0.0
html5lib==0.99999
Pillow==2.8.0
pytz==2015.4
six==1.9.0
tzlocal==1.1.3
wheel==0.24.0
